+++ src/main/java/com/ssg/meowwms/dto/storage/QrcodeDTO.java
@@ -1,4 +1,4 @@
-package com.ssg.meowwms.domain;
+package com.ssg.meowwms.dto.storage;
 
 import lombok.*;
 
@@ -10,7 +10,7 @@ import java.util.Date;
 @AllArgsConstructor
 @NoArgsConstructor
 @ToString
-public class QrcodeVO {
+public class QrcodeDTO {
     private int id;
     private int productId;
     private Blob barcodeData;
