+++ src/main/java/com/ssg/meowwms/mapper/storage/ProductMapper.java
@@ -1,6 +1,6 @@
 package com.ssg.meowwms.mapper.storage;
 
-import com.ssg.meowwms.domain.ProductVO;
+import com.ssg.meowwms.domain.storage.ProductVO;
 import org.apache.ibatis.annotations.Mapper;
 
 @Mapper
