+++ src/main/resources/mappers/storage/QrMapper.xml
@@ -4,7 +4,7 @@
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
 <mapper namespace="com.ssg.meowwms.mapper.storage.QrMapper">
-<!--    <insert id="saveQrBarcode" parameterType="com.ssg.meowwms.domain.QrcodeVO">-->
+<!--    <insert id="saveQrBarcode" parameterType="com.ssg.meowwms.domain.storage.QrcodeVO">-->
 <!--        INSERT INTO Qr_Barcode (product_id, barcode_Data, creation_Date)-->
 <!--        VALUES (#{productId}, #{barcodeData, jdbcType=BLOB}, #{creationDate})-->
 <!--    </insert>-->
