+++ src/main/resources/mappers/storage/StockMovementMapper.xml
@@ -3,7 +3,7 @@
 <mapper namespace="com.ssg.meowwms.mapper.storage.StockMovementMapper">
 
     <!-- insertStock -->
-    <insert id="insertStock" parameterType="com.ssg.meowwms.domain.StockMovementVO">
+    <insert id="insertStock" parameterType="com.ssg.meowwms.domain.storage.StockMovementVO">
         INSERT INTO Stock_Movement (product_id, user_id, status_code, request_datetime, approved_datetime)
         VALUES (#{productId}, #{userId}, #{statusCode}, #{requestDatetime}, #{approvedDatetime})
     </insert>
@@ -16,7 +16,7 @@
     </update>
 
     <!-- updateStockMovement -->
-    <update id="updateStockMovement" parameterType="com.ssg.meowwms.domain.StockMovementVO">
+    <update id="updateStockMovement" parameterType="com.ssg.meowwms.domain.storage.StockMovementVO">
         UPDATE stock_movement
         SET
             product_id = #{productId},
@@ -28,17 +28,17 @@
     </update>
 
     <!-- selectAllStockMovements -->
-    <select id="selectAllStockMovements" resultType="com.ssg.meowwms.domain.StockMovementVO">
+    <select id="selectAllStockMovements" resultType="com.ssg.meowwms.domain.storage.StockMovementVO">
         SELECT * FROM stock_movement
     </select>
 
     <!-- selectMovementByStatus -->
-    <select id="selectMovementByStatus" resultType="com.ssg.meowwms.domain.StockMovementVO" parameterType="string">
+    <select id="selectMovementByStatus" resultType="com.ssg.meowwms.domain.storage.StockMovementVO" parameterType="string">
         SELECT * FROM stock_movement WHERE status_code = #{statusCode}
     </select>
 
     <!-- selectStockMovementsById -->
-    <select id="selectStockMovementsById" resultType="com.ssg.meowwms.domain.StockMovementVO" parameterType="int">
+    <select id="selectStockMovementsById" resultType="com.ssg.meowwms.domain.storage.StockMovementVO" parameterType="int">
         SELECT * FROM stock_movement WHERE product_id = #{productId}
     </select>
